Fix typo in SpongeBob's entry
I didn't write this; one of the author's of vo.codes did. I don't know if this was supposed to be a joke, but they said SpongeBob lives in a "sponge" under the sea instead of a pineapple. Pretty disturbing if you ask me. I'm changing it to how I remember the lyrics went.